ai_service/license_verifier/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.9
pytesseract==0.3.13
Pillow==10.4.0
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
